python3-kornia: Downgrade to 0.6.12
In 0.7.x, kornia_rs became a mandatory dependency without
mentioning it in the release notes. Unfortunately, kornia_rs
needs the maturin build backend which is not available in
mickledore and nanbield.

In 0.6.x, kornia_rs was an optional dependency so downgrade
to the latest 0.6.12 in that series.

Also set PE = "1" so package repositories can upgrade this package.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>