python3-pytorch: Upgrade to 2.4.0
Simplified the long SRC_URI list. Some gitsm:// URIs work fine
but the one for kineto still doesn't.

Removed -DUSE_TBB=ON and the tbb build dependency from the recipe
as it's not used by PyTorch 2.4.0 anymore.

Removed a patch that doesn't apply to the new version.
Adapted a patch to the new version.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>