conf/layer.conf: clang-revival-layer to recommends
The clang depended recipe - python3-llvmlite is already located in a
dynamic-layers config based on meta-clang. Though as meta-clang-revival
depend on meta-clang, it is not used.

With the move to LAYERRECOMMNDS instead of LAYERDEPENDS, it is possible
to build components from meta-python-ai without including both
meta-clang-revival and meta-clang.

As the current version of python3-llvmlite depend on clang15 from
meta-clang-revival it could be considered to move it to a dynamic layer
for meta-clang-revival. Though as the python3-llvmlite is updated to a
clang version from meta-clang, it would move back and forth - so just
keep it in meta-clang.

Signed-off-by: Claus Stovgaard <claus.stovgaard@prevas.dk>